pyproject.toml
[project]
name = "otto"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Nwosu-Ihueze", email = "nwosunneoma@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "chardet>=5.2.0",
    "modin>=0.36.0",
    "pandas>=2.3.2",
    "pyarrow>=21.0.0",
    "python-magic>=0.4.27",
    "tiktoken>=0.11.0",
    "torch>=2.8.0",
    "transformers>=4.56.2",
    "unstructured>=0.18.15",
    "werkzeug>=3.1.3",
]

[project.scripts]
otto = "otto:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
